Moved unit tests to tests folder - Deleted unittest folder
